phi_service/requirements.txt
phidata
ollama
flask
duckduckgo-search